in zone editor, group records in record sets, and show their propagation status & history
in the zone editor, records that have previous versions still out there in dns
caches have an orange box in front of them. if negative lookup results are
still cached, they'll have a yellow box. if both, it's an orange box with
yellow border.

there is a new button "History" for each record set. it shows the complete
propagation history. that too shows boxes in front of earlier versions to
indicate if they could still be in dns caches.

negative ttl's and relevant wildcard records are taken into accounts. dname
records are not (could be done in the future). if other record types are
relevant, they aren't handled at the moment...