in web interface, after an error trying to add records, do properly refresh the zone/records when adding succeeds a second or later attempt
The fix removes some error handling. The "await check client" calls already
show a popup with an error. Rejecting the promise while the edit popup is still
opens essentially raises an error at the caller, causing it to not refresh the
zone.  We now only resolve the popup either when closed/canceled, or on
success.