handle updates to just the SOA record better
before this commit, we would declare a zone as unchanged when only the
non-serial field(s) of the soa record changed. e.g. a new primary ns or
mbox, etc. we now explicitly look for that change.

before, editing the soa record through the web interface always appeared
to change. the upstream dns server may have made the change, but we
wouldn't sync it back. only when changing a non-SOA record would we
start seeing the change (with a soa record that changed to a newer
serial again).

keep in mind upstream dns servers may require you change the serial when
you're updating any other fields of the soa record. knot does.

this always ensure changing the SOA record through the web interface
always causes a "set" in the underlying "provider", never a
"delete"+"add" (which is normally triggered when the provider doesn't
have record id's). there's a good chance providers don't like deletion
and addition of soa records...